Refactor URL parsing and shortcode addition
This commit refactors the code to parse the URL using `parse_url()` function and adds shortcodes based on the GET parameters. The `wf_solar_install_map()` function is modified to include a style tag for Google Maps.

Tried uncommenting some shit to get the map working on the final results page